requirements.txt
asgiref==3.8.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
Django==5.2.3
django-allauth==65.9.0
django-formtools==2.5.1
django-jazzmin==3.0.1
django-otp==1.6.0
django-phonenumber-field==8.1.0
django-tinymce==4.1.0
django-two-factor-auth==1.17.0
idna==3.10
phonenumbers==9.0.8
pillow==11.3.0
pycparser==2.22
PyJWT==2.10.1
pypng==0.20220715.0
qrcode==7.4.2
requests==2.32.4
sqlparse==0.5.3
typing_extensions==4.14.0
urllib3==2.5.0
